pulled out base request module for examples in ruby
