+++ ruby/post_ruby.rb
@@ -1,50 +1,12 @@
-require 'dotenv'
-require 'json'
-require 'jwt'
-require 'rest-client'
+require_relative './request.rb'
 
-Dotenv.load
-
-access_key = ENV['ACCESS_KEY']
-secret_key = ENV['SECRET_KEY']
-base_url   = ENV['API_URL']
-
-body = { companies: [
+params = { companies: [
 	{ 
-		name: 'Sample Z', 
+		name: 'Sample abc', 
 		entity_type: 'Limited Liability Company', 
 		jurisdictions: ['Maine', 'Washington']
 	}
 ]}.to_json
 
-payload = {
-  path: '/companies',
-  content: Digest::SHA2.hexdigest(body)
-}
-
-headers = { access_key: access_key }
-
-token = JWT.encode(
-    payload,
-    secret_key,
-    'HS256',
-    headers
-)
-
-begin
-  res = RestClient::Request.execute(
-    method: :post,
-    url: base_url + '/companies',
-    payload: body,
-    headers: {
-      authorization: "Bearer #{token}",
-      content_type: :json
-    }
-  )
-  
-  pp JSON(res.body)
-rescue
-
-  pp JSON($!.response)
-end
+BaseRequestRoute.request(:post, '/companies', body: params)
